pdf-resources: move resource tests out of module
Move resource parsing regressions into a dedicated integration test
target so they exercise the crate through its public API.

Route the XObject cache coverage through Resources::read while
preserving generated content stream ID assertions.
